Move keypoint to ops.
